greentea/mysql: Add `mysql_ping` wrapper
`int mysql_ping(MYSQL *mysql)`

@# Description
Checks whether the connection to the server is working. If it has gone
down, an automatic reconnection is attempted.

This function can be used by clients that remain idle for a long while,
to check whether the server has closed the connection and reconnect if
necessary.

@# Return Values
Zero if the server is alive. Non-zero if an error occurred.

@# Errors
`CR_COMMANDS_OUT_OF_SYNC`
    Commands were executed in an improper order.

`CR_SERVER_GONE_ERROR`
    The MySQL server has gone away.

`CR_UNKNOWN_ERROR`
    An unknown error occurred.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>