greentea/mysql: Fix `prepareLen()` retval
Note that this function allocates a buffer for `bind_` with `calloc()`.
Before this commit, if the `calloc()` fails, it returns `nullptr`. Then
the caller will check for MySQL error after that.

However, this situation is not MySQL error, thus we need to inform the
caller that we run out of memory.

Fix this, return `ERR_PTR(-ENOMEM)` if we fail on `calloc()`.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>