greentea/tgvisd: Prevent race condition
The assignment to scraperThread_ outside the lambda might've not
happened yet we the thread's PC enters the lambda. Thus
`this->scraperThread_` may be NULL. Let's make sure the assignment
has been done before we run the scraper work by busy waiting in a
loop.

It should be fine to spinning here as the period must be very short.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>