greentea/tgvisd: Fix use-after-free bug
There is a use-after-free bug caused by race condition. This part is
the wrong thing to do:
```
  tasks_[kwrk_id].is_used = true;
  tasks_[kwrk_id].tw = std::move(*tw);
```
Note that on `_run_kworker()` we have this check:
```
  twl = kwrk_pool->getTW(kwrk_id);
  if (!twl->is_used)
          return;
```
If @twl->is_used is true but the @twl->tw still contains a dangling
pointer, then it can lead to use-after-free in the `_run_kworker()`.
Don't do that, make sure when @twl->is_used is set to true, the
@twl->tw has been changed with a valid pointer that `std::move(*tw);`
gives.

Fix for this is simple, make sure when `is_used` is set to true, the
`tw` does not contain a dangling pointer, so we can reorder the
assignment like this:
```
  tasks_[kwrk_id].tw = std::move(*tw);
  tasks_[kwrk_id].is_used = true;
```

Extra small changes:
  - Mark the `is_used` as volatile.
  - Reset the @tw.chat on `putTaskWork()` to reduce the work on
   `submitTaskWork()`.

Fixes: 71174b375e52b6513f6800c7ff6cb4dfc6d1b08e ("greetea/tgvisd: Implement scraper kworker")
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>