greentea/tgvisd: Add join queue to handle tgvkwrk-%u exit
When the `KWorker::runThreadPool()` is exiting, it can't call
`thread->join()` itself because the task to be waited is itself.

The workaround for this was using `thread->detach()` and free
the associated resource with it. But it can't guarantee the
task has exited right after the resource gets freed, so it leads
to race condition with the `KWorker::getThPool()`. This race
condition caused UAF bug as well. ASAN caught that and had me
debugged this for couple of hours pondering the call traces!

In that case, we can't hold `thPoolLock_` neither. Because we
are going to exit, nobody can release the lock when it exits.
The master kworker may do, but nobody notifies it.

^ We can't do that again!

Better solution for this is:
 - Add a queue work for master kworker to clean these up.
 - Call `thread->join()` from the master kworker based on
   the queue work.
 - After `thread->join()`, we can guarantee that the task
   has exited, so it resolves all the problems went here.

It consumes extra space for `std::queue<uint32_t>` and
`std::mutex`. But I don't see that as a problem.

This is the best way I can think of right now. If you think
there is a better workaround for this, you can email me and
argue with your technical standpoint.

Fixes: 5ab179d5a7922acf168dd1b634a5e14890deadd8 ("greentea/tgvisd: Implement several td methods on kworker")
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>