greentea/tgvisd: Td: Fix thread cancelation bug
Alviro reports that he found a segmentation fault on the lambda that
the `query_sync_callback()` creates.

The story behind the bug comes from `Td::send_query_sync()`. It
creates a callback that tdlib is supposed to process. The problem with
that callback is that, it holds a pointer to a local variable on
`Td::send_query_sync()`'s stack.

When we get a cancelation from the main thread, `Td::send_query_sync()`
returns without even care whether the callback still holds a pointer to
its local variable. Once we return, the local variable is invalid, its
lifetime ends, results in undefined behavior from the callback.

Fix this by making a promise. Wrap the argument with a struct that
lives on the heap. Then add an extra variable `uint8_t cleaner`. This
variabe is protected with a mutex, if the value is `QSD_CLEAN_FROM_SQS`,
we do the clean up from `Td::send_query_sync()`. Otherwise we do the
clean up from the callback.

Link: https://t.me/TeaInside/45071
Reported-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>
Tested-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>