greentea/tgvisd: Logger: Add sender abstractions
This adds sender abstractions for database normalization. The main
issue this patch addresses is that the sender may not be a user, it
can be a group or channel. So we normalize it here.

Currently, there are 2 types of sender:
  1. User
  2. Chat (this includes group and channel)

This patch only creates new files for the abstractions without
changing the functionality. We will use it in the later patches.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Link: https://t.me/TeaInside/47415
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>