greentea/tgvisd: Logger: Add chat abstractions
This adds chat abstractions for database normalization. The main issue
this patch addresses is that the chat has different kind of types.
Currently there are 4 types of chat:
  1. chatTypeBasicGroup
  2. chatTypeSupergroup
  3. chatTypeSecret
  4. chatTypePrivate

For 1 and 2, they stand for group and channel. For 3 and 4, they stand
for private chat.

This patch only creates new files for the abstractions without
changing the functionality. We will use it in the later patches.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Link: https://t.me/TeaInside/47416
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>