greentea/tgvisd: Move `getDbPool()` call to Message class
This commit fixes a broken logging design that was introduced in commit
8d4025f4951304fb2 ("greentea/tgvisd: Logger: Add getDbPool in foundation
classes").

Do not call `getDbPool()` (instance of `KWorker`) in the ChatFoundation
or SenderFoundation class. Because their destructor will put the DB
pool and we do not actually need two DB pools for this operation.

Consider the following:
```
    pk_chat_id_ = m_chat_->getPK();
    if (unlikely(!pk_chat_id_))
        goto out;

    pk_sender_id_ = m_sender_->getPK();
    if (unlikely(!pk_sender_id_))
        goto out;
```
After those two `getPK()` calls, as long as `m_chat_` and `m_sender_`
still live, they will be holding DB pool instance individually, each
needs to be put on destruction.

That is only the partial craziness, we still have another problem
here. Recall that we still need a DB pool again in the Message class
to save the message, thus we will need to at least hold 3 DB pools
just to save a single message (including chat and sender data), which
clearly does not make any sense!

Why are we wasting so many DB pools just for this small operation!?

So what is the fix for this?
We can heavily reuse this DB pool from the Message class, now we take
the DB pool from KWorker in there, then pass it to ChatFoundation and
SenderFoundation. End result, we only need 1 DB pool to save a single
message.

Cc: Iskandar S (Telegram) <@Viro_SSFS>
Cc: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Fixes: 8d4025f4951304fb2 ("greentea/tgvisd: Logger: Add getDbPool in foundation classes")
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>