greentea/tgvisd: Message: Handle message forward info
This adds message forward info handler. The message forward info
handler is not perfect at this point, but we save everything that
the extra meta data has with "extra" column by serializing the
object into a string. Someday when we need to normalize it, we can
arrange a code to deserialize this extra string.

While in there, update the DDL to support this message forward info
handler.

Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>