greentea/tgvisd: Wrap arguments to save function with a ctx struct
We will need many things to be accessed along with the logger
development. This wraps arguments to the save message function with
a context struct so we can easily add more stuff to be processed
later via this context struct.

Reasoning:
If we pass all of them via argument, we will have many arguments to
be passed in, that is painful, better use a context struct and pass
the pointer to that struct.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>